Bootstrap CLA signers file for automated checks
This file will hold the list of currently-authorized JanusGraph contributors for
automated verification by tools and processes, while we're still using the PDF
method for processing ICLA / CCLA signatures.

In the future, we may not need this file once the signature process is automated
and migrated to a fully-electronic system end-to-end.

Signed-off-by: Misha Brukman <mbrukman@google.com>