Fix @myhrvold's name to match GitHub account
Exact name and email in the CLA must match that of the patch:
https://github.com/JanusGraph/legal/pull/41.patch
or the CLA bot is not happy and marks the commit, and hence the PR, as
non-conformant to the signed CLAs on file.

Longer term, the more elegant way to handle this in the `CLA_SIGNERS.yaml` file
will be to add alternate names and email addresses for everyone to make this
process easier.

Signed-off-by: Misha Brukman <mbrukman@google.com>