Add graceful support for handling empty fields
Outputs a clear message to show which field(s) are empty/unspecified to
allow easily tracking down an error in the config (found while
attempting to make a change and forgetting to set a field).

This change also adds some basic tests to enable making safer changes to
the validation logic in the future, as adding a test is now as simple as
adding a new test method.

Signed-off-by: Misha Brukman <mbrukman@google.com>