Convert code to functional style
Instead of directly outputting errors to stderr, simply return a tuple of
(valid, errors) and let the caller handle how to output it.

This also cleans up the error reporting when running tests, and will allow us to
test error messages if we choose to do so in the future.

Signed-off-by: Misha Brukman <mbrukman@google.com>