Remove Hortonworks CCLA as it merged with Cloudera
Hortonworks merged with Cloudera, but the name of the combined company
is "Cloudera" so Hortonworks (the brand name) has been retired.
Additionally, https://hortonworks.com redirects to https://cloudera.com,
which also suggests that the old name is no longer being used.

Thus, we need to remove the Hortonworks CCLA as it is no longer in
force, and if folks from the combined Cloudera/Hortonworks entity want
to contribute to JanusGraph, they will need to have Cloudera sign a new
CCLA, which might as well be done via the new electronic LF EasyCLA
process rather than the old-style process.

Signed-off-by: Misha Brukman <mbrukman@google.com>